requirements.txt
allure-pytest==2.14.2
pytest==8.3.5
requests==2.32.3
selenium==4.33.0

